fix(import): Clean up external support, fixes CLI
