--- packages/concerto-cli/test1.json
@@ -1,1 +1,0 @@
-{"$class":"concerto.metamodel.Model","namespace":"test1","imports":[],"declarations":[{"$class":"concerto.metamodel.ConceptDeclaration","name":"Foo1","isAbstract":false,"properties":[{"$class":"concerto.metamodel.IntegerProperty","name":"a","isArray":false,"isOptional":false,"location":{"$class":"concerto.metamodel.Range","start":{"offset":260,"line":7,"column":2,"$class":"concerto.metamodel.Position"},"end":{"offset":285,"line":8,"column":1,"$class":"concerto.metamodel.Position"}},"validator":{"$class":"concerto.metamodel.IntegerDomainValidator","lower":0,"upper":10}}],"location":{"$class":"concerto.metamodel.Range","start":{"offset":244,"line":6,"column":1,"$class":"concerto.metamodel.Position"},"end":{"offset":286,"line":8,"column":2,"$class":"concerto.metamodel.Position"}}}]}
