+++ packages/concerto-core/api.txt
@@ -218,7 +218,7 @@ class ModelManager {
    + DecoratorFactory[] getDecoratorFactories() 
    + void addDecoratorFactory(DecoratorFactory) 
    + boolean derivesFrom(string,string) 
-   + void getMetaModel() 
+   + void getAst() 
 }
 class SecurityException extends BaseException {
    + void constructor(string)
