+++ packages/concerto-core/changelog.txt
@@ -24,7 +24,7 @@
 # Note that the latest public API is documented using JSDocs and is available in api.txt.
 #
 
-Version 1.2.2 {f45f585f2937493f94e71478aa61d703} 2021-11-24
+Version 1.2.2 {859dc5db772867c4591ed1fd70b8f840} 2021-11-24
 - Remove custom instanceof and add methods to check runtime type
 - Remove support for Node 12
 - Generate Typescript definitions from JSDoc
