+++ packages/concerto-cto/test/external.js
@@ -22,7 +22,7 @@ const should = chai.should();
 chai.use(require('chai-things'));
 chai.use(require('chai-as-promised'));
 
-const { resolveExternal } = require('../lib/parserMain');
+const { resolveExternal } = require('../lib/external');
 
 /**
  * Get the name and content of all cto files
